redis与spring security的初步结合

+++ src/main/java/com/easyauth/controller/admin/UserController.java
@@ -50,14 +50,14 @@ public class UserController {
      */
     @Operation(summary = "切换用户状态")
     @PutMapping("/status")
-    public Result<String> switchStatus(Long id, Long status) {
+    public Result<String> switchStatus(Integer id, Integer status) {
         userService.switchStatus(id, status);
         return Result.success();
     }
 
     @Operation(summary = "用户列表", description = "current必须提供，size默认为10")
     @GetMapping("/list")
-    public Result<Page<UserVO>> getList(Long current, @RequestParam(required = false, defaultValue = "10") Long size) {
+    public Result<Page<UserVO>> getList(Integer current, @RequestParam(required = false, defaultValue = "10") Integer size) {
         return Result.success(userService.getList(current, size));
     }
