+++ src/main/java/com/easyauth/domain/DTO/UserPageQueryDTO.java
@@ -6,9 +6,9 @@ import java.util.List;
 
 @Data
 public class UserPageQueryDTO {
-    Long current;
-    Long size = 10L;
+    Integer current;
+    Integer size = 10;
     String username;
     String email;
-    List<Long> roleIds;
+    List<Integer> roleIds;
 }
